fixed column width
The mobile first setting was messing with the width of my columns in the reading and toplayer views. This was mostly because it was not set up properly: the width does not change between `main.css` and `desktop.css`.

I just removed the setting for now, because I don't have time to fix it. And also, I may need to rethink the whole approach since I am working with multiple `<div class="row">` elements now.